Changed bootstrap logic to prefer files from the repository. Renamed repository files to .sample to avoid misconceptions.

+++ application/physical/scripts/sdcardsetup.sh
@@ -267,6 +267,7 @@ if [ ! -f ./application/physical/systemd/etc/systemd/network/70_rpi-3-b_usbports
     <trim />NamePolicy=mac
     <trim />#Name=lan4
 EOF
+  $_SUDO cp ./tmp/root/etc/systemd/network/70_rpi-3-b_usbports.link ./application/physical/systemd/etc/systemd/network/70_rpi-3-b_usbports.link
 else
   $_SUDO cp ./application/physical/systemd/etc/systemd/network/70_rpi-3-b_usbports.link ./tmp/root/etc/systemd/network/70_rpi-3-b_usbports.link
 fi
@@ -281,6 +282,7 @@ sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/systemd/network/70_rp
   <trim />DHCP=ipv4
   <trim />IPv6AcceptRouterAdvertisements=0
 EOF
+  $_SUDO cp ./tmp/root/etc/systemd/network/70_rpi-3-b_usbports.network ./application/physical/systemd/etc/systemd/network/70_rpi-3-b_usbports.network
 else
   $_SUDO cp ./application/physical/systemd/etc/systemd/network/70_rpi-3-b_usbports.network ./tmp/root/etc/systemd/network/70_rpi-3-b_usbports.network
 fi
@@ -294,6 +296,7 @@ if [ ! -f ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernet
     <trim />[Link]
     <trim />Name=lan0
 EOF
+  $_SUDO cp ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.link ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.link
 else
   $_SUDO cp ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.link ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.link
 fi
@@ -324,6 +327,7 @@ if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then
 
   for d in etc/salt/minion.d etc/salt/master.d srv/salt srv/pillar srv/pillar/users srv/pillar/network; do
     if [ ! -d ./tmp/root/${d} ]; then mkdir -p ./tmp/root/${d}; fi
+    if [ ! -d ./application/physical/saltstack/${d} ]; then mkdir -p ./application/physical/saltstack/${d}; fi
   done
 
   #TODO: Change sudoers logic to only grant ALL rights when an individual password is supplied
@@ -339,62 +343,57 @@ if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then
 EOF
 
   #master saltstack salt-master ext_pillar
-  if [ ! -z ${__GITREMOTEORIGINURL} ]; then #should always match
-    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/99_ext_pillar.conf > /dev/null
-      <trim />ext_pillar:
-      <trim />  - git:
-      <trim />    - ${__GITREMOTEORIGINBRANCH} ${__GITREMOTEORIGINURL}:
-      <trim />      - root: application/physical/saltstack/srv/pillar
+  if [ ! -f ./application/physical/saltstack/etc/salt/master.d/90_ext_pillar.conf ]; then
+    if [ ! -z ${__GITREMOTEORIGINURL} ]; then #repeated for clarity/safety/...
+      sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/90_ext_pillar.conf > /dev/null
+        <trim />ext_pillar:
+        <trim />  - git:
+        <trim />    - ${__GITREMOTEORIGINBRANCH} ${__GITREMOTEORIGINURL}:
+        <trim />      - root: application/physical/saltstack/srv/pillar
 EOF
-    #echo "/etc/salt/master.d/99_ext_pillar.conf"
-    #cat ./tmp/root/etc/salt/master.d/99_ext_pillar.conf
-  elif [ -f ./application/physical/saltstack/etc/master.d/90_ext_pillar.conf ]; then
-    $_SUDO cp ./application/physical/saltstack/etc/master.d/90_ext_pillar.conf ./tmp/root/etc/salt/master.d/90_ext_pillar.conf
-    #echo "/etc/salt/master.d/90_ext_pillar.conf"
-    #cat ./tmp/root/etc/salt/master.d/90_ext_pillar.conf
+      $_SUDO cp ./tmp/root/etc/salt/master.d/90_ext_pillar.conf ./application/physical/saltstack/etc/salt/master.d/90_ext_pillar.conf
+    fi
+  else
+    $_SUDO cp ./application/physical/saltstack/etc/salt/master.d/90_ext_pillar.conf ./tmp/root/etc/salt/master.d/90_ext_pillar.conf
   fi
 
   #master saltstack salt-master gitfs_remotes
-  if [ ! -z ${__GITREMOTEORIGINURL} ]; then #repeated for clarity
-  sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/99_gitfs_remotes.conf > /dev/null
-    <trim />gitfs_remotes:
-    <trim />  - ${__GITREMOTEORIGINURL}:
-    <trim />    - root: application/physical/saltstack/srv/salt
-    <trim />    - base: ${__GITREMOTEORIGINBRANCH}
-    <trim />  - ${__GITREMOTEORIGINURL}:
-    <trim />    - name: dnsmasq_formula
-    <trim />    - root: application/physical/vendor/saltstack-formulas/dnsmasq_formula
-    <trim />    - base: ${__GITREMOTEORIGINBRANCH}
-    <trim />  - ${__GITREMOTEORIGINURL}:
-    <trim />    - name: ntp_formula
-    <trim />    - root: application/physical/vendor/saltstack-formulas/ntp_formula
-    <trim />    - base: ${__GITREMOTEORIGINBRANCH}
-    <trim />  - ${__GITREMOTEORIGINURL}:
-    <trim />    - name: users_formula
-    <trim />    - root: application/physical/vendor/saltstack-formulas/users_formula
-    <trim />    - base: ${__GITREMOTEORIGINBRANCH}
+  if [ ! -f ./application/physical/saltstack/etc/salt/master.d/90_gitfs_remotes.conf ]; then
+    if [ ! -z ${__GITREMOTEORIGINURL} ]; then #repeated for clarity/safety/...
+      sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/90_gitfs_remotes.conf > /dev/null
+        <trim />gitfs_remotes:
+        <trim />  - ${__GITREMOTEORIGINURL}:
+        <trim />    - root: application/physical/saltstack/srv/salt
+        <trim />    - base: ${__GITREMOTEORIGINBRANCH}
+        <trim />  - ${__GITREMOTEORIGINURL}:
+        <trim />    - name: dnsmasq_formula
+        <trim />    - root: application/physical/vendor/saltstack-formulas/dnsmasq_formula
+        <trim />    - base: ${__GITREMOTEORIGINBRANCH}
+        <trim />  - ${__GITREMOTEORIGINURL}:
+        <trim />    - name: ntp_formula
+        <trim />    - root: application/physical/vendor/saltstack-formulas/ntp_formula
+        <trim />    - base: ${__GITREMOTEORIGINBRANCH}
+        <trim />  - ${__GITREMOTEORIGINURL}:
+        <trim />    - name: users_formula
+        <trim />    - root: application/physical/vendor/saltstack-formulas/users_formula
+        <trim />    - base: ${__GITREMOTEORIGINBRANCH}
 EOF
-    #echo "/etc/salt/master.d/99_gitfs_remotes.conf"
-    #cat ./tmp/root/etc/salt/master.d/99_gitfs_remotes.conf
-  elif [ -f ./application/physical/saltstack/etc/master.d/90_gitfs_remotes.conf ]; then
-    $_SUDO cp ./application/physical/saltstack/etc/master.d/90_gitfs_remotes.conf ./tmp/root/etc/salt/master.d/90_gitfs_remotes.conf
-    #echo "/etc/salt/master.d/90_gitfs_remotes.conf"
-    #cat ./tmp/root/etc/salt/master.d/90_gitfs_remotes.conf
+    $_SUDO cp ./tmp/root/etc/salt/master.d/90_gitfs_remotes.conf ./application/physical/saltstack/etc/salt/master.d/90_gitfs_remotes.conf
+    fi
+  else
+    $_SUDO cp ./application/physical/saltstack/etc/salt/master.d/90_gitfs_remotes.conf ./tmp/root/etc/salt/master.d/90_gitfs_remotes.conf
   fi
 
   #master saltstack salt-master fileserver_backends
-  if [ ! -f ./application/physical/saltstack/etc/master.d/90_fileserver_backends.conf ]; then
-    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/90_fileserver_backends.conf > /dev/null
+  if [ ! -f ./application/physical/saltstack/etc/salt/master.d/90_fileserver_backend.conf ]; then
+    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/salt/master.d/90_fileserver_backend.conf > /dev/null
       <trim />fileserver_backend:
       <trim />  - roots
       <trim />  - git
 EOF
-    #echo "/etc/salt/master.d/90_fileserver_backends.conf"
-    #cat ./tmp/root/etc/salt/master.d/90_fileserver_backends.conf
+    $_SUDO cp ./tmp/root/etc/salt/master.d/90_fileserver_backend.conf ./application/physical/saltstack/etc/salt/master.d/90_fileserver_backend.conf
   else
-    $_SUDO cp ./application/physical/saltstack/etc/master.d/90_fileserver_backends.conf ./tmp/root/etc/salt/master.d/90_fileserver_backends.conf
-    #echo "/etc/salt/master.d/90_fileserver_backends.conf"
-    #cat ./tmp/root/etc/salt/master.d/90_fileserver_backends.conf
+    $_SUDO cp ./application/physical/saltstack/etc/salt/master.d/90_fileserver_backend.conf ./tmp/root/etc/salt/master.d/90_fileserver_backend.conf
   fi
 
   #master saltstack minions grains
@@ -408,10 +407,11 @@ EOF
     <trim />    - ntp-server
     <trim />#    - proxy-cache
     <trim />#    - network-router
-    <trim />#    - power-switch
+    <trim />    - power-button
 EOF
 
   #master saltstack pillar
+  #I /believe/ that the pillar system is more fragile so I do not change the
   sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/pillar/top.sls > /dev/null
     <trim />base:
     <trim />  '*':
@@ -433,43 +433,52 @@ EOF
       <trim />include:
       <trim />  - ntp/ng
 EOF
-    #echo "/etc/salt/master.d/90_fileserver_backends.conf"
-    #cat ./tmp/root/srv/salt/ntp-server.sls
+    $_SUDO cp ./tmp/root/srv/salt/ntp-server.sls ./application/physical/saltstack/srv/salt/ntp-server.sls
   else
     $_SUDO cp ./application/physical/saltstack/srv/salt/ntp-server.sls ./tmp/root/srv/salt/ntp-server.sls
-    #echo "/srv/salt/ntp-server.sls"
-    #cat ./tmp/root/srv/salt/ntp-server.sls
   fi
-  sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/systemd/timesyncd.conf > /dev/null
-    <trim />[Time]
-    <trim />NTP=${__MASTERIP}
-    <trim />FallbackNTP=0.arch.pool.ntp.org 1.arch.pool.ntp.org 2.arch.pool.ntp.org 3.arch.pool.ntp.org
+
+  if [ ! -f ./application/physical/systemd/etc/systemd/timesyncd.conf ]; then #not in a cloned/forked repo
+    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/systemd/timesyncd.conf > /dev/null
+      <trim />[Time]
+      <trim />NTP=${__MASTERIP}
+      <trim />FallbackNTP=0.arch.pool.ntp.org 1.arch.pool.ntp.org 2.arch.pool.ntp.org 3.arch.pool.ntp.org
 EOF
+    $_SUDO cp ./tmp/root/etc/systemd/timesyncd.conf ./application/physical/systemd/etc/systemd/timesyncd.conf
+  else
+    $_SUDO cp ./application/physical/systemd/etc/systemd/timesyncd.conf ./tmp/root/etc/systemd/timesyncd.conf
+  fi
 
 
   #master saltstack pillar ntp-server
   #TODO: Test reliance on helotism file
   #FIXME: Breaks on nonexisting file /srv/pillar/helotism.sls
-  sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/pillar/ntp-server.sls > /dev/null
-    <trim />{% from 'helotism.sls' import vars %}
-    <trim />ntp:
-    <trim />  ng:
-    <trim />    settings:
-    <trim />      ntpd: True
-    <trim />      ntp_conf:
-    <trim />        server: ['127.127.1.1', '0.de.pool.ntp.org', '1.de.pool.ntp.org', '2.de.pool.ntp.org', '3.de.pool.ntp.org']
-    <trim />        fudge: ['127.127.1.1 stratum 12']
-    <trim />        interface: ['listen lan0', 'listen lo' ]
-    <trim />        driftfile: ['/var/lib/ntp/ntp.drift']
-    <trim />{% if vars.helotism.__NETWORKSEGMENT is defined and vars.helotism.__NETWORKPREFIX is defined %}
-    <trim />        restrict: ['default nomodify nopeer noquery', '127.0.0.1', '::1', '{{ vars.helotism.__NETWORKPREFIX }} mask {{ vars.helotism.__NETWORKSEGMENT }} nomodify nopeer notrap']
-    <trim />{% else %}
-    <trim />        restrict: ['default nomodify nopeer noquery', '127.0.0.1', '::1', '${__NETWORKPREFIX} mask ${__NETWORKSEGMENT} nomodify  nopeer notrap']
-    <trim />{% endif %}
+  if [ ! -f ./application/physical/saltstack/srv/pillar/ntp-server.sls ]; then
+    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/pillar/ntp-server.sls > /dev/null
+      <trim />{% from 'helotism.sls' import vars %}
+      <trim />ntp:
+      <trim />  ng:
+      <trim />    settings:
+      <trim />      ntpd: True
+      <trim />      ntp_conf:
+      <trim />        server: ['127.127.1.1', '0.de.pool.ntp.org', '1.de.pool.ntp.org', '2.de.pool.ntp.org', '3.de.pool.ntp.org']
+      <trim />        fudge: ['127.127.1.1 stratum 12']
+      <trim />        interface: ['listen lan0', 'listen lo' ]
+      <trim />        driftfile: ['/var/lib/ntp/ntp.drift']
+      <trim />{% if vars.helotism.__NETWORKSEGMENT is defined and vars.helotism.__NETWORKPREFIX is defined %}
+      <trim />        restrict: ['default nomodify nopeer noquery', '127.0.0.1', '::1', '{{ vars.helotism.__NETWORKPREFIX }} mask {{ vars.helotism.__NETWORKSEGMENT }} nomodify nopeer notrap']
+      <trim />{% else %}
+      <trim />        restrict: ['default nomodify nopeer noquery', '127.0.0.1', '::1', '${__NETWORKPREFIX} mask ${__NETWORKSEGMENT} nomodify  nopeer notrap']
+      <trim />{% endif %}
 EOF
+    $_SUDO cp ./tmp/root/srv/pillar/ntp-server.sls ./application/physical/saltstack/srv/pillar/ntp-server.sls
+  else
+     $_SUDO cp ./application/physical/saltstack/srv/pillar/ntp-server.sls ./tmp/root/srv/pillar/ntp-server.sls
+  fi
 
 
   #master saltstack pillar dhcp-server dns-server
+  if [ ! -f ./application/physical/saltstack/srv/pillar/dhcp-server.sls ]; then
   sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/pillar/dhcp-server.sls > /dev/null
     <trim />dnsmasq:
     <trim />  dnsmasq_conf: salt://dnsmasq/files/dnsmasq.conf
@@ -505,48 +514,66 @@ EOF
     <trim />    auth-zone:
     <trim />      - ${__FQDNNAME},${__NETWORKSEGMENTCIDR}
 EOF
+    $_SUDO cp ./tmp/root/srv/pillar/dhcp-server.sls ./application/physical/saltstack/srv/pillar/dhcp-server.sls
+  else
+     $_SUDO cp ./application/physical/saltstack/srv/pillar/dhcp-server.sls ./tmp/root/srv/pillar/dhcp-server.sls
+  fi
 
   #master saltstack salt
   #$_SUDO tee ./tmp/root/srv/salt/top.sls > /dev/null <<EOF
-  sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/salt/top.sls > /dev/null
-    <trim />base:
-    <trim />  '*':
-    <trim />    - helotism
-    <trim />    - common
-    <trim />
-    <trim />  'roles:dhcp-server':
-    <trim />    - match: grain
-    <trim />    - dhcp-server
-    <trim />
-    <trim />  'roles:ntp-server':
-    <trim />    - match: grain
-    <trim />    - ntp-server
-    <trim />
-    <trim />  'roles:power-button':
-    <trim />    - match: grain
-    <trim />    - power-button
+  if [ ! -f ./application/physical/saltstack/srv/salt/top.sls ]; then
+    sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/srv/salt/top.sls > /dev/null
+      <trim />base:
+      <trim />  '*':
+      <trim />    - helotism
+      <trim />    - common
+      <trim />
+      <trim />  'roles:dhcp-server':
+      <trim />    - match: grain
+      <trim />    - dhcp-server
+      <trim />
+      <trim />  'roles:ntp-server':
+      <trim />    - match: grain
+      <trim />    - ntp-server
+      <trim />
+      <trim />  'roles:power-button':
+      <trim />    - match: grain
+      <trim />    - power-button
 EOF
+    $_SUDO cp ./tmp/root/srv/salt/top.sls ./application/physical/saltstack/srv/salt/top.sls
+  else
+     $_SUDO cp ./application/physical/saltstack/srv/salt/top.sls ./tmp/root/srv/salt/top.sls
+  fi
 
   if [ -f ./tmp/srv_pillar_helotism.sls ]; then cp ./tmp/srv_pillar_helotism.sls ./tmp/root/srv/pillar/helotism.sls ; fi
 
-  #master systemd network
+if [ ! -f ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.router.network ]; then #not in a cloned/forked repo
   sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.router.network > /dev/null
-    <trim />[Match]
-    <trim />Name=lan0
-
-    <trim />[Network]
-    <trim />Address=${__NETWORKSEGMENTCIDR}
-    <trim />IPForward=ipv4
-    <trim />IPMasquerade=yes
-    <trim />IPv6AcceptRouterAdvertisements=0
+      <trim />[Match]
+      <trim />Name=lan0
+  
+      <trim />[Network]
+      <trim />Address=${__NETWORKSEGMENTCIDR}
+      <trim />IPForward=ipv4
+      <trim />IPMasquerade=yes
+      <trim />IPv6AcceptRouterAdvertisements=0
 EOF
+  $_SUDO cp ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.router.network ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.router.network
+else
+  $_SUDO cp ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.router.network ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.router.network
+fi
 
   #master systemd resolved
+if [ ! -f ./application/physical/systemd/etc/systemd/resolved.conf ]; then
   sed 's/^[ ]*<trim \/>//' <<EOF | $_SUDO tee ./tmp/root/etc/systemd/resolved.conf > /dev/null
     <trim />[Resolve]
     <trim />DNS=${__MASTERIP}
     <trim />Domains=${__FQDNNAME}
 EOF
+  $_SUDO cp ./tmp/root/etc/systemd/resolved.conf ./application/physical/systemd/etc/systemd/resolved.conf
+else
+  $_SUDO cp ./application/physical/systemd/etc/systemd/resolved.conf ./tmp/root/etc/systemd/resolved.conf
+fi
 
 ######################################################################
 #/**
@@ -568,6 +595,7 @@ Name=lan0
 DHCP=ipv4
 IPv6AcceptRouterAdvertisements=0
 EOF
+    $_SUDO cp ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.nonrouter.network ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.nonrouter.network
   else
     $_SUDO cp ./application/physical/systemd/etc/systemd/network/20_rpi-3-b_ethernetport.nonrouter.network ./tmp/root/etc/systemd/network/20_rpi-3-b_ethernetport.nonrouter.network
   fi
@@ -576,7 +604,9 @@ fi #end if else master=hostname
 
 ######################################################################
 #/**
-#  * This place intentionally left blank
+#  * This is a poor man's firstboot setup.
+#  * 
+#  * @ToDo: When it turns out stable, automate the process (there is no user input anyways) 
 #  *
 #  */
 
@@ -599,7 +629,7 @@ if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then tmp="${tmp} -M"; fi
 _todolive="pacman -Syu --noconfirm";
 if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then _todolive=" ${_todolive}; pacman -S --noconfirm python2-pygit2"; fi
 _todolive=" ${_todolive}; curl -o bootstrap_salt.sh -L https://bootstrap.saltstack.com --silent -k; sleep 2; $_SUDO sh bootstrap_salt.sh -U -i ${__HOSTNAME} ${tmp} git v2016.3.0";
-if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then _todolive=" ${_todolive}; sleep 60; salt-key -A -y; echo "Sleeping 60 seconds to settle down salt."; sleep 60; salt '${__MASTERHOSTNAME}' state.apply dnsmasq; salt '${__MASTERHOSTNAME}' state.apply ntp-server" salt '${__MASTERHOSTNAME}' state.apply common; salt '${__MASTERHOSTNAME}' state.apply power-button; fi
+if [ "$__HOSTNAME" = "$__MASTERHOSTNAME" ]; then _todolive=" ${_todolive}; sleep 60; salt-key -A -y; echo 'Sleeping 60 seconds to settle down salt.'; sleep 60; salt '${__MASTERHOSTNAME}' state.apply dnsmasq; salt '${__MASTERHOSTNAME}' state.apply ntp-server; salt '${__MASTERHOSTNAME}' state.apply common; salt '${__MASTERHOSTNAME}' state.apply power-button"; fi
 _todolive=" ${_todolive}; timedatectl set-ntp true;"
 
 echo ${_todolive} >> ./tmp/root_root_.bash_history
@@ -636,4 +666,4 @@ echo "Login as root, followed by:"
 echo ${_todolive}
 echo ""
 echo "This was written into the bash history of the root user. Login as root and press the arrow up key."
-sleep 2
+sleep 6
